requirements.txt
confluent-kafka
protobuf
grpcio
grpcio-tools